Add the modifier helpers to another file
