+++ tests/EventEnvelopesT.cpp
@@ -4,7 +4,7 @@
 // license. You should have receive a LICENSE.md file along with the code.
 // If not, contact the sfizz maintainers at https://github.com/sfztools/sfizz
 
-#include "sfizz/SfzHelpers.h"
+#include "sfizz/ModifierHelpers.h"
 #include "sfizz/Buffer.h"
 #include "catch2/catch.hpp"
 #include <absl/algorithm/container.h>
